Update `test_harness` for spec v0.4.0
